+++ python/cuxfilter/datatile.py
@@ -34,6 +34,8 @@ class DataTile:
         """
         if self.passive_chart.chart_type == "datasize_indicator":
             return self._calc_data_tile_for_size(data)
+        elif self.passive_chart.chart_type == "3d_choropleth":
+            return self._calc_3d_choropleth_data_tile(data.copy())
         if self.dimensions == 2:
             return self._calc_2d_data_tile(data.copy())
 
@@ -63,3 +65,33 @@ class DataTile:
             cumsum=self.cumsum,
             return_format=self.dtype,
         )
+
+    def _calc_3d_choropleth_data_tile(self, data):
+        """
+        calc multiple data tiles for color and elevation agg for 3d choropleth
+        """
+        ret_datatile = {}
+        print(self.cumsum)
+        self.passive_chart.y = self.passive_chart.color_column
+        ret_datatile[
+            self.passive_chart.color_column
+        ] = gpu_datatile.calc_data_tile(
+            data,
+            self.active_chart,
+            self.passive_chart,
+            self.passive_chart.color_aggregate_fn,
+            cumsum=self.cumsum,
+            return_format=self.dtype,
+        )
+        self.passive_chart.y = self.passive_chart.elevation_column
+        ret_datatile[
+            self.passive_chart.elevation_column
+        ] = gpu_datatile.calc_data_tile(
+            data,
+            self.active_chart,
+            self.passive_chart,
+            self.passive_chart.elevation_aggregate_fn,
+            cumsum=self.cumsum,
+            return_format=self.dtype,
+        )
+        return ret_datatile