+++ CHANGELOG.md
@@ -13,3 +13,8 @@
 - PR #78 Fix gpuCI GPU build script
 - PR #83 Fix conda upload
 - PR #84 Updated Docs and Readme with conda nightly install instructions for cuxfilter version 0.12
+- PR #86 Implemented #79 - cudatashader replaced by datashader(>=0.9) with cudf & dask_cudf support
+- PR #87 Implemented jupyter-server-proxy as discussed in #73
+- PR #89 Fixed headless chrome sandbox for dashboard preview feature issue mentioned in #88
+- PR #90 Implemented deck-gl_bokeh plugin and integrated with cuxfilter with layout and theme options
+  and added full support for deck.gl/polygon layer